+++ tests/acceptance/transitions/test_track_cmds.py
@@ -17,6 +17,7 @@ from tests.utils import remove_subscriptions, setup_subscriptions
 TRACKING_POSITION_THRESHOLD_ERROR_DEG = 0.05
 INIT_AZ = -250
 INIT_EL = 70
+POINTING_TOLERANCE_DEG = 0.1
 
 
 @pytest.fixture
@@ -25,10 +26,12 @@ def slew_dish_to_init(event_store_class, dish_manager_proxy):
     main_event_store = event_store_class()
     band_event_store = event_store_class()
     achieved_pointing_event_store = event_store_class()
+    pointing_state_event_store = event_store_class()
     attr_cb_mapping = {
         "dishMode": main_event_store,
         "configuredBand": band_event_store,
         "achievedPointing": achieved_pointing_event_store,
+        "pointingstate": pointing_state_event_store,
     }
     subscriptions = setup_subscriptions(dish_manager_proxy, attr_cb_mapping)
 
@@ -39,22 +42,17 @@ def slew_dish_to_init(event_store_class, dish_manager_proxy):
     # Await auto transition to OPERATE following band config
     main_event_store.wait_for_value(DishMode.OPERATE, timeout=10, proxy=dish_manager_proxy)
 
-    current_az, current_el = dish_manager_proxy.achievedPointing[1:]
-    estimate_slew_duration = max(abs(INIT_EL - current_el), (abs(INIT_AZ - current_az) / 3))
     dish_manager_proxy.Slew([INIT_AZ, INIT_EL])
 
-    # wait until no updates
-    data_points = achieved_pointing_event_store.get_queue_values(
-        timeout=estimate_slew_duration + 10
-    )
-    # timeout return empty list
-    assert data_points
-    # returned data is an array of tuple consisting of attribute name and value
-    last_az_el = data_points[-1][1]
-    # check last az and el received and compare with reference
-    achieved_az, achieved_el = last_az_el[1], last_az_el[2]
-    assert achieved_az == pytest.approx(INIT_AZ)
-    assert achieved_el == pytest.approx(INIT_EL)
+    def target_reached_test(pointing_event_val: list[float]) -> bool:
+        """Return whether we got to the target."""
+        return pointing_event_val[1] == pytest.approx(
+            INIT_AZ, abs=POINTING_TOLERANCE_DEG
+        ) and pointing_event_val[2] == pytest.approx(INIT_EL, abs=POINTING_TOLERANCE_DEG)
+
+    achieved_pointing_event_store.wait_for_condition(target_reached_test, timeout=120)
+    pointing_state_event_store.wait_for_value(PointingState.READY, timeout=120)
+
     remove_subscriptions(subscriptions)
 
     yield
@@ -123,8 +121,7 @@ def test_track_and_track_stop_cmds(
 
     [[_], [unique_id]] = dish_manager_proxy.Track()
     result_event_store.wait_for_command_id(unique_id, timeout=8)
-    pointing_state_event_store.wait_for_value(PointingState.SLEW, timeout=6)
-    pointing_state_event_store.wait_for_value(PointingState.TRACK, timeout=6)
+    pointing_state_event_store.wait_for_value(PointingState.TRACK, timeout=60)
 
     expected_progress_updates = [
         "Fanned out commands: DS.Track",
@@ -154,7 +151,7 @@ def test_track_and_track_stop_cmds(
         )
 
     achieved_pointing_event_store.clear_queue()
-    achieved_pointing_event_store.wait_for_condition(check_final_points_reached, timeout=20)
+    achieved_pointing_event_store.wait_for_condition(check_final_points_reached, timeout=60)
 
     # Call TrackStop on DishManager
     [[_], [unique_id]] = dish_manager_proxy.TrackStop()